% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_win.R
\name{check_win}
\alias{check_win}
\title{Check which positions will win}
\usage{
check_win(board, target = 1, n = 4)
}
\arguments{
\item{board}{The board matrix}

\item{target}{Integer: What player id do you want.}

\item{n}{Integer:  Connect-n (usually 4)}
}
\value{
A vector of will win flags in each slot
}
\description{
Check which positions will win
}
\examples{
set.seed(12)
x <- matrix(sample(1:2, size = 24, replace = TRUE), nrow = 4)
x <- apply(x, 2, function(y) {
  i=sample(4, size=1)-1
  y[1:i] <- NA
  y})

}
