% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-dougbot2000.R
\docType{data}
\name{dougbot2000}
\alias{dougbot2000}
\title{The unbeatable dougbot2000 c4 player!}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dougbot2000
}
\description{
The unbeatable dougbot2000 c4 player!
}
\section{Public Members and Methods}{


\code{name}: character. The name of the bot.

\code{make_move(board, lastMove)}: Get the next move, supply the current board matrix and the last move

\code{rejected(board, lastMove)}: Get the next move, supply the current board matrix and the last move
}

\examples{
dougbot2000$new(name="dougbotkill")

}
\keyword{datasets}
