% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-dougbot2001.R
\docType{data}
\name{dougbot2001}
\alias{dougbot2001}
\title{The unbeatable dougbot2001 c4 player!}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dougbot2001
}
\description{
The unbeatable dougbot2001 c4 player!
}
\section{Public Members and Methods}{


\code{name}: character. The name of the bot.

\code{make_move(board, lastMove)}: Get the next move, supply the current board matrix and the last move

\code{rejected(board, lastMove)}: Get the next move, supply the current board matrix and the last move
}

\examples{
dougbot2001$new(name="dougbotkill", player = 1)

}
\keyword{datasets}
